Always log response from Azkaban during upload
We currently only log the response from Azkaban when it's a 200 status code, which makes it hard to troubleshoot what went wrong when it's a non-200 status (we just print the status line and the response code). This moves the logging of the response above where we check if it was an error so that it's printed regardless.